# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zorven REST API
  version: "1.0.0"
  description: |
    Zorven'in programatik REST API'si. Tünellerinizi, istemcilerinizi,
    hostname'lerinizi ve IP kurallarınızı yönetmenizi; istek loglarını ve
    abonelik/kullanım verinizi okumanızı sağlar.

    ## Kimlik doğrulama
    Panelde **API Anahtarları** bölümünden bir token oluşturun ve her istekte
    `Authorization` başlığında gönderin:

    ```
    Authorization: Bearer zrv_api_xxxxxxxx...
    ```

    API erişimi **Pro ve üzeri** planlarda kullanılabilir.

    ## Kapsamlar (scopes)
    Her token bir dizi kapsam taşır (`kaynak:aksiyon`):
    `clients:read`, `clients:write`, `tunnels:read`, `tunnels:write`,
    `hostnames:read`, `hostnames:write`, `ip_allowlist:read`,
    `ip_allowlist:write`, `analytics:read`. `write`, aynı kaynağın `read`'ini
    de kapsar. Yetersiz kapsam → `403 insufficient_scope`.

    ## Sayfalama
    Liste uçları `?limit=` (varsayılan 50, maks 200) ve `?cursor=` kabul eder.
    Yanıt gövdesi düz bir dizidir; devam bilgisi başlıklarda döner:
    `X-Next-Cursor`, `X-Has-More` ve RFC5988 `Link: <...>; rel="next"`.
    limit/cursor verilmezse tüm liste döner.

    ## Hatalar
    Tüm hatalar ortak biçimdedir:
    ```json
    { "error": { "code": "insufficient_scope", "message": "..." } }
    ```

    ## Hız sınırı
    Programatik istekler kiracı başına ~30 istek/sn (burst 60) ile sınırlıdır.
    Yanıtlarda `X-RateLimit-Limit` ve `X-RateLimit-Remaining` döner; aşımda
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /tunnels:
    get:
      tags: [Tunnels]
      summary: Tünelleri listele
      description: "Kapsam: `tunnels:read`"
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Tünel listesi
          headers:
            X-Next-Cursor: { $ref: "#/components/headers/XNextCursor" }
            X-Has-More: { $ref: "#/components/headers/XHasMore" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tunnel" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Tunnels]
      summary: Tünel oluştur
      description: "Kapsam: `tunnels:write`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, client_id, target]
              properties:
                name: { type: string, description: "Kısa ad (tek DNS etiketi)", example: "api" }
                client_id: { type: string, example: "cli_a1b2c3" }
                target: { type: string, example: "http://localhost:8000" }
                hostname_id: { type: string, description: "Opsiyonel mevcut hostname'e bağla" }
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tunnel" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /tunnels/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Tunnels]
      summary: Tünel getir
      description: "Kapsam: `tunnels:read`"
      responses:
        "200":
          description: Tünel
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tunnel" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Tunnels]
      summary: Tünel güncelle
      description: "Kapsam: `tunnels:write`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target: { type: string, example: "http://localhost:9000" }
                enabled: { type: boolean }
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tunnel" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      tags: [Tunnels]
      summary: Tünel sil
      description: "Kapsam: `tunnels:write`"
      responses:
        "204": { description: Silindi }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: zrv_api_...
      description: "Panelde oluşturulan API token'ı. `Authorization: Bearer zrv_api_...`"
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 50, maximum: 200 }
      description: Sayfa boyutu (opt-in sayfalama)
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Önceki yanıtın X-Next-Cursor değeri
  headers:
    XNextCursor:
      description: Sonraki sayfa cursor'ı (varsa)
      schema: { type: string }
    XHasMore:
      description: Daha fazla kayıt var mı
      schema: { type: boolean }
  responses:
...
    Forbidden:
      description: Yetersiz kapsam, plan kısıtı veya erişim yok
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Kaynak bulunamadı
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Çakışma (ör. hostname zaten alınmış)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Geçersiz girdi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: insufficient_scope }
            message: { type: string }
      required: [error]
...
    Tunnel:
      type: object
      properties:
        id: { type: string, example: "tun_x9y8z7" }
        client_id: { type: string, example: "cli_a1b2c3" }
        target: { type: string, example: "http://localhost:8000" }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        hostnames:
          type: array
          items: { $ref: "#/components/schemas/Hostname" }
    Hostname:
      type: object
      properties:
        id: { type: string, example: "hn_abc123" }
        tunnel_id: { type: string }
        fqdn: { type: string, example: "api.zorven.app" }
        type: { type: string, enum: [global, scoped, legacy, custom], description: "global = kısa ad, scoped = kiracı adı" }
        verified: { type: boolean }
        created_at: { type: string, format: date-time }
...